split out loading module
