+++ src/components/loading/indicator.tsx
@@ -1,44 +1,10 @@
 import { useClient } from "@/hooks/client";
-import { tw } from "@/utils";
 import { Store, useStore } from "@/utils/store";
 import type { PropsWithChildren } from "react";
 import { useEffect, useId, useRef } from "react";
 import { createPortal } from "react-dom";
-import { Transition } from "./animation/transition";
-
-const DURATION = 1000;
-const DOT_COUNT = 12;
-const DOTS = Array.from({ length: DOT_COUNT });
-type SpinnerProps = { className?: string };
-export function Spinner({ className }: SpinnerProps) {
-  return (
-    <svg
-      viewBox="0 0 24 24"
-      className={className}
-    >
-      {DOTS.map((_, i) => (
-        <circle
-          key={i}
-          cx="12"
-          cy="2"
-          r="0"
-          transform={`rotate(${360 * (i / DOT_COUNT)})`}
-          className={tw("origin-center", i % 2 ? "fill-app-green-500" : "fill-app-green-500/30")}
-        >
-          <animate
-            attributeName="r"
-            dur={`${DURATION}ms`}
-            begin={`${i * (DURATION / DOT_COUNT)}ms`}
-            keySplines="0.2 0.2 0.4 0.8;0.2 0.2 0.4 0.8;0.2 0.2 0.4 0.8"
-            values="0;2;0;0"
-            calcMode="spline"
-            repeatCount="indefinite"
-          />
-        </circle>
-      ))}
-    </svg>
-  );
-}
+import { Transition } from "../animation/transition";
+import { Spinner } from "./spinner";
 
 class LoadingStore extends Store<boolean> {
   private stack: string[] = [];
