+++ src/pages/product/[id].tsx
@@ -1,8 +1,8 @@
-import { Spinner } from "@/components/Loading";
 import { SortDialog, useParseSort } from "@/components/Search";
 import { Star, UserPic } from "@/components/UI";
 import { Layout } from "@/components/layout";
 import { InfiniteScroll } from "@/components/list/infinite-scroll";
+import { Spinner } from "@/components/loading/spinner";
 import {
   CategoryButton,
   ConsIcon,
