+++ src/pages/profile/index.tsx
@@ -2,7 +2,7 @@ import { Layout } from "@/components/layout";
 import { HeaderLink } from "@/components/layout/header";
 import { InfiniteScroll } from "@/components/list/infinite-scroll";
 import { Card, NoResults } from "@/components/list/product";
-import { Spinner } from "@/components/Loading";
+import { Spinner } from "@/components/loading/spinner";
 import { HeaderSearchBar, SEARCH_QUERY_KEY, SortDialog, useParseSort } from "@/components/Search";
 import { Star, UserPic } from "@/components/UI";
 import { fetchNextPage, minutesToMs } from "@/utils";
