+++ src/pages/review/[id]/index.tsx
@@ -1,7 +1,7 @@
-import { Layout } from "@/components/layout";
-import { useSetLoadingIndicator } from "@/components/Loading";
 import { toast } from "@/components/Toast";
 import { Button, DialogOverlay, Star, UrlDialogRoot } from "@/components/UI";
+import { Layout } from "@/components/layout";
+import { useSetLoadingIndicator } from "@/components/loading/indicator";
 import {
   BarcodeTitle,
   CategoryButton,
